ai_bot/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
torch
transformers
peft
trl
bitsandbytes
accelerate
datasets
langchain
langchain-community
sentence-transformers
faiss-cpu
unstructured
python-multipart